Hardcoded endpoint version to v1.0.
This is to strongly associate client library with specific version of
cloud API. Also fixed following issues:

Issue with trailing '/' in BaseUrl.

Issue in PDF endpoint which was preventing ErrorJSON to recieve proper
value in case of StatusCode is other than 200 and in other endpoints
only StatusCode 400 was being respected.